+++ src/features/index.ts
@@ -2,7 +2,19 @@ export * from './commands/handlers/meta/help'
 export * from './commands/handlers/lifecycle/chatSpawn'
 export * from './commands/handlers/behaviors/behaviors'
 export * from './commands/handlers/behaviors/breakBlock'
-export * from './commands/handlers/youAreMine'
+export * from './commands/handlers/inventory/equipment-swap'
+export * from './commands/handlers/inventory/inventory-swap'
+export * from './commands/handlers/inventory/mainhand-swap'
+export * from './commands/handlers/inventory/offhand-swap'
+export * from './commands/handlers/inventory/recycle'
+export * from './commands/handlers/lifecycle/chatSpawn'
+export * from './commands/handlers/lifecycle/disconnect'
+export * from './commands/handlers/lifecycle/respawn'
+export * from './commands/handlers/meta/help'
+export * from './commands/handlers/meta/list'
+export * from './commands/handlers/meta/location'
+export * from './commands/handlers/meta/rename'
+export * from './commands/handlers/meta/time'
 export * from './task'
 export * from './gui'
 export * from './killedBySimPlayer'
